requirements.txt
flask
gunicorn
pyjwt
